requirements.txt
amqp==2.5.2
asgiref==3.2.3
billiard==3.6.1.0
celery==4.4.0
dj-database-url==0.5.0
dj-email-url==0.2.0
Django==3.0.2
djangorestframework==3.11.0
environs==7.1.0
importlib-metadata==1.4.0
kombu==4.6.7
marshmallow==3.3.0
more-itertools==8.1.0
psycopg2-binary==2.8.4
python-dotenv==0.10.3
pytz==2019.3
sqlparse==0.3.0
vine==1.3.0
zipp==2.0.1
